병합전 커밋
