+++ src/main/java/com/flab/daitso/service/OrderService.java
@@ -1,18 +1,28 @@
 package com.flab.daitso.service;
 
-/*import com.flab.daitso.dto.order.OrderDto;
-import com.flab.daitso.mapper.OrderMapper;
+import com.flab.daitso.dto.order.OrderDto;
+//import com.flab.daitso.mapper.OrderMapper;
+import com.flab.daitso.dto.order.OrderRequestDto;
+import com.flab.daitso.dto.product.ProductDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class OrderService {
 
-    @Autowired
-    OrderMapper orderMapper;
+    //@Autowired
+    //OrderMapper orderMapper;
 
-    public void orderNormal(OrderDto orderDto){
+    public OrderDto orderNormal(ProductDto productDto, OrderRequestDto orderRequestDto){
         //charge nothing more
+        double productPrice = productDto.getPrice();
+        if (orderRequestDto.getCoupon() != null){
+            productPrice *= orderRequestDto.getCoupon().getAmount();
+        }
+
+
+        return null;
+
     }
 
     public void orderRocket(OrderDto orderDto){
@@ -22,4 +32,4 @@ public class OrderService {
     public void orderWow(OrderDto orderDto){
         //charge more than above
     }
-}*/
+}
